+++ CHANGELOG.md
@@ -3,6 +3,7 @@ T## [Unreleased]
 
 * Add async SDL events handling (#166)
 * Implement relative mouse inputs
+* Add a setting to warp the pointer to the center of the screen after each frame
 
 ### Changed